folio/rails combined require
If you're in a rails project, you can just require 'folio/rails' to get
ActiveRecord pagination support, ActionController rescue handlers, and
an ActionView will_paginate view helper.

requires that the will_paginate gem be installed in your project (won't
pull it in for you, and will raise a descriptive method if it's not
there).

ActionView stuff will come in a later commit, but I wanted to setup the
requires for it now with the rest.

Change-Id: Id7d730cb27a582c1be3a5a00220f95ac0d5479b2