immediately instantiate on paginate/page for AR
will_paginate 3.0 returns a relation from paginate(...). this allows
chaining of further relation modifiers. when the relation is finally
instantiated a WillPaginate::Collection is returned. however:

 (1) this is inconsistent with the will_paginate 2.3 behavior where
     paginate(...) immediately returns an instantiation of the scope as
     a WillPaginate::Collection obect.

 (2) this can cause unintentional errors with rails3's interpretation
     of, for example, map!. relation.map! { ... } on rails3 silently
     does nothing (I consider this a rails3 bug, but I'm just going to
     work around it for now). typically, you wouldn't call .map! on a
     relation; you'd call map, or you'd instantiate it first. but if the
     caller of paginate doesn't know it didn't instantiate, they may try
     to call map! and not achieve the desired effect.

the need to chain relation modifiers after paginate is rare, if it even
exists, and is easily worked around, while the detriments are subtle and
nasty. so we just break with WillPaginate and instantiate the (page of
the) relation immediately on paginate.

Change-Id: I7933c069a51932dbfd7b5f4ea08664a3fec7d2ac