fix auto-counting with grouping
consider Item.group(:some_attr).paginate. since we're not providing
total_entries to the paginate call, it would try and use
Item.group(:some_attr).count as the total_entries. but that's going to
return a hash -- e.g. {'value1' => 12, 'value2' => 17, 'value3' => 9}.

when rows are actually returned, there's will be one row each per
'value1', 'value2', etc. so we're really concerned with the number of
unique values for the group by clause. select precisely that to use as
the total_entries before passing it along to Folio proper.

Change-Id: Ia94569b27891bf37f141bbddde59ba1000fe3309